feat(document-model-libs): update debounce logic
