+++ packages/document-model-libs/editors/document-model-2/components/code-editors/index.ts
@@ -1,0 +1,2 @@
+export * from "./graphql-editor";
+export * from "./json-editor";
