+++ packages/document-model-libs/editors/document-model-2/components/index.ts
@@ -1,6 +1,6 @@
 export * from "./form";
 export * from "./label";
-export * from "./graphql-editor";
+export * from "./code-editors/graphql-editor";
 export * from "./module-form";
 export * from "./input";
-export * from "./json-editor";
+export * from "./code-editors/json-editor";
