+++ packages/document-model-libs/editors/document-model-2/constants/documents.ts
@@ -18,3 +18,5 @@ export const specialDocIds = {
   global: "global",
   local: "local",
 } as const;
+
+export const updateTimeout = 5000;
